ci: update Makefile and add copyright headers
Updated the Makefile to include a new variable `UVCMD` and modified the
`generate` and `check` targets to use it. Added copyright headers to
`code/prow.py` and `pipeline-prow.yaml`. Improved the `check` target to
use a temporary file for comparison and provide a more detailed diff
output if changes are needed.

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>