feat: Add check runs status before merging PR
This commit introduces a check for the status of check runs
before merging a pull request. The PR will only be merged if
all required checks have passed. If any checks have failed or
are still pending, a comment will be posted on the PR with a
table of the failed or pending checks and the merge will be
aborted. This change ensures that only PRs with successful
checks are merged.

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>