fix: handle admin/write user merge logic
This commit updates the PRHandler class to handle cases where an admin
or write user can merge directly if the LGTM threshold is 1. It also
ensures that their merge command counts as a vote if not already counted
when the threshold is greater than 1. Additionally, it adds missing
newlines in the argument parser for better readability. The help message
in messages.py is updated to reflect these changes.

Fixes #35

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>